refactor(): Expand and parse special forms, not macros
Rather than trying to add a special case for every conditional macro
possible, instead call `Macro.expand` on the steps of a `pipet` and
parse the resulting special forms (ie, `cond`, `case`, and `with`). This
greatly simplifies the implementation, and also allows for supporting
macros that users write that expand down to those special forms.